(#4282) - refactor merge.js, remove hyphens
