+++ extras/ajax.js
@@ -1,4 +1,4 @@
 'use strict';
 
 // allow external plugins to require('pouchdb/extras/ajax')
-module.exports = require('../lib/deps/ajax/ajax-core');
+module.exports = require('../lib/deps/ajax/ajaxCore');
