+++ lib/adapter.js
@@ -1,7 +1,6 @@
 "use strict";
 
 var utils = require('./utils');
-var merge = require('./merge');
 var errors = require('./deps/errors');
 var EventEmitter = require('events').EventEmitter;
 var upsert = require('./deps/upsert');
@@ -9,6 +8,10 @@ var Changes = require('./changes');
 var Promise = utils.Promise;
 var isDeleted = require('./deps/docs/isDeleted');
 var isLocalId = require('./deps/docs/isLocalId');
+var traverseRevTree = require('./deps/merge/traverseRevTree');
+var collectLeaves = require('./deps/merge/collectLeaves');
+var rootToLeaf = require('./deps/merge/rootToLeaf');
+var collectConflicts = require('./deps/merge/collectConflicts');
 
 /*
  * A generic pouch adapter
@@ -70,7 +73,7 @@ function compareByIdThenRev(a, b) {
 function computeHeight(revs) {
   var height = {};
   var edges = [];
-  merge.traverseRevTree(revs, function (isLeaf, pos, id, prnt) {
+  traverseRevTree(revs, function (isLeaf, pos, id, prnt) {
     var rev = pos + "-" + id;
     if (isLeaf) {
       height[rev] = 0;
@@ -347,7 +350,7 @@ AbstractPouchDB.prototype.revsDiff =
   function processDoc(id, rev_tree) {
     // Is this fast enough? Maybe we should switch to a set simulated by a map
     var missingForId = req[id].slice(0);
-    merge.traverseRevTree(rev_tree, function (isLeaf, pos, revHash, ctx,
+    traverseRevTree(rev_tree, function (isLeaf, pos, revHash, ctx,
       opts) {
         var rev = pos + '-' + revHash;
         var idx = missingForId.indexOf(rev);
@@ -409,7 +412,7 @@ AbstractPouchDB.prototype.compactDocument =
       }
     });
 
-    merge.traverseRevTree(revTree, function (isLeaf, pos, revHash, ctx, opts) {
+    traverseRevTree(revTree, function (isLeaf, pos, revHash, ctx, opts) {
       var rev = pos + '-' + revHash;
       if (opts.status === 'available' && candidates.indexOf(rev) !== -1) {
         revs.push(rev);
@@ -515,7 +518,7 @@ AbstractPouchDB.prototype.get =
         if (err) {
           return callback(err);
         }
-        leaves = merge.collectLeaves(rev_tree).map(function (leaf) {
+        leaves = collectLeaves(rev_tree).map(function (leaf) {
           return leaf.rev;
         });
         finishOpenRevs();
@@ -550,7 +553,7 @@ AbstractPouchDB.prototype.get =
     var ctx = result.ctx;
 
     if (opts.conflicts) {
-      var conflicts = merge.collectConflicts(metadata);
+      var conflicts = collectConflicts(metadata);
       if (conflicts.length) {
         doc._conflicts = conflicts;
       }
@@ -561,7 +564,7 @@ AbstractPouchDB.prototype.get =
     }
 
     if (opts.revs || opts.revs_info) {
-      var paths = merge.rootToLeaf(metadata.rev_tree);
+      var paths = rootToLeaf(metadata.rev_tree);
       var path = arrayFirst(paths, function (arr) {
         return arr.ids.map(function (x) { return x.id; })
           .indexOf(doc._rev.split('-')[1]) !== -1;
