+++ lib/adapters/idb/allDocs.js
@@ -1,9 +1,9 @@
 'use strict';
 
-var merge = require('../../merge');
 var errors = require('../../deps/errors');
 var idbUtils = require('./utils');
 var idbConstants = require('./constants');
+var collectConflicts = require('../../deps/merge/collectConflicts');
 
 var ATTACH_STORE = idbConstants.ATTACH_STORE;
 var BY_SEQ_STORE = idbConstants.BY_SEQ_STORE;
@@ -98,7 +98,7 @@ function idbAllDocs(opts, api, idb, callback) {
       docIdRevIndex.get(key).onsuccess =  function onGetDoc(e) {
         row.doc = decodeDoc(e.target.result);
         if (opts.conflicts) {
-          row.doc._conflicts = merge.collectConflicts(metadata);
+          row.doc._conflicts = collectConflicts(metadata);
         }
         fetchAttachmentsIfNecessary(row.doc, opts, txn);
       };
