+++ lib/adapters/idb/index.js
@@ -1,7 +1,6 @@
 'use strict';
 
 var utils = require('../../utils');
-var merge = require('../../merge');
 var isDeleted = require('../../deps/docs/isDeleted');
 var isLocalId = require('../../deps/docs/isLocalId');
 var errors = require('../../deps/errors');
@@ -11,6 +10,8 @@ var idbBulkDocs = require('./bulkDocs');
 var idbAllDocs = require('./allDocs');
 var checkBlobSupport = require('./blobSupport');
 var hasLocalStorage = require('../../deps/env/hasLocalStorage');
+var calculateWinningRev = require('../../deps/merge/winningRev');
+var traverseRevTree = require('../../deps/merge/traverseRevTree');
 
 var ADAPTER_VERSION = idbConstants.ADAPTER_VERSION;
 var ATTACH_AND_SEQ_STORE = idbConstants.ATTACH_AND_SEQ_STORE;
@@ -117,7 +118,7 @@ function init(api, opts, callback) {
         var metadata = cursor.value;
         var docId = metadata.id;
         var local = isLocalId(docId);
-        var rev = merge.winningRev(metadata);
+        var rev = calculateWinningRev(metadata);
         if (local) {
           var docIdRev = docId + "::" + rev;
           // remove all seq entries
@@ -230,7 +231,8 @@ function init(api, opts, callback) {
       }
       var metadata = decodeMetadataCompat(cursor.value);
 
-      metadata.winningRev = metadata.winningRev || merge.winningRev(metadata);
+      metadata.winningRev = metadata.winningRev ||
+        calculateWinningRev(metadata);
 
       function fetchMetadataSeq() {
         // metadata.seq was added post-3.2.0, so if it's missing,
@@ -623,7 +625,7 @@ function init(api, opts, callback) {
 
     docStore.get(docId).onsuccess = function (event) {
       var metadata = decodeMetadata(event.target.result);
-      merge.traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
+      traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
                                                          revHash, ctx, opts) {
         var rev = pos + '-' + revHash;
         if (revs.indexOf(rev) !== -1) {
