+++ lib/adapters/leveldb/index.js
@@ -15,8 +15,10 @@ function requireLeveldown() {
 requireLeveldown();
 
 var errors = require('../../deps/errors');
-var merge = require('../../merge');
 var utils = require('../../utils');
+var calculateWinningRev = require('../../deps/merge/winningRev');
+var traverseRevTree = require('../../deps/merge/traverseRevTree');
+var collectConflicts = require('../../deps/merge/collectConflicts');
 var isDeleted = require('../../deps/docs/isDeleted');
 var isLocalId = require('../../deps/docs/isLocalId');
 var processDocs = require('../../deps/docs/processDocs');
@@ -65,7 +67,7 @@ var safeJsonEncoding = {
 // in newer versions of PouchDB, they are cached on the metadata
 function getWinningRev(metadata) {
   return 'winningRev' in metadata ?
-    metadata.winningRev : merge.winningRev(metadata);
+    metadata.winningRev : calculateWinningRev(metadata);
 }
 
 function getIsDeleted(metadata, winningRev) {
@@ -864,7 +866,7 @@ function LevelPouch(opts, callback) {
             doc.doc = data;
             doc.doc._rev = doc.value.rev;
             if (opts.conflicts) {
-              doc.doc._conflicts = merge.collectConflicts(metadata);
+              doc.doc._conflicts = collectConflicts(metadata);
             }
             for (var att in doc.doc._attachments) {
               if (doc.doc._attachments.hasOwnProperty(att)) {
@@ -1130,7 +1132,7 @@ function LevelPouch(opts, callback) {
         return callback(err);
       }
       var seqs = metadata.rev_map; // map from rev to seq
-      merge.traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
+      traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
                                                          revHash, ctx, opts) {
         var rev = pos + '-' + revHash;
         if (revs.indexOf(rev) !== -1) {
