+++ lib/adapters/websql/index.js
@@ -1,13 +1,14 @@
 'use strict';
 
 var utils = require('../../utils');
-var merge = require('../../merge');
 var isDeleted = require('../../deps/docs/isDeleted');
 var isLocalId = require('../../deps/docs/isLocalId');
 var errors = require('../../deps/errors');
 var parseHexString = require('../../deps/parseHex');
 var binStringToBlob = require('../../deps/binary/binaryStringToBlobOrBuffer');
 var hasLocalStorage = require('../../deps/env/hasLocalStorage');
+var collectConflicts = require('../../deps/merge/collectConflicts');
+var traverseRevTree = require('../../deps/merge/traverseRevTree');
 
 var websqlConstants = require('./constants');
 var websqlUtils = require('./utils');
@@ -684,7 +685,7 @@ function WebSqlPouch(opts, callback) {
               doc.doc = data;
               doc.doc._rev = winningRev;
               if (opts.conflicts) {
-                doc.doc._conflicts = merge.collectConflicts(metadata);
+                doc.doc._conflicts = collectConflicts(metadata);
               }
               fetchAttachmentsIfNecessary(doc.doc, opts, api, tx);
             }
@@ -876,7 +877,7 @@ function WebSqlPouch(opts, callback) {
       var sql = 'SELECT json AS metadata FROM ' + DOC_STORE + ' WHERE id = ?';
       tx.executeSql(sql, [docId], function (tx, result) {
         var metadata = utils.safeJsonParse(result.rows.item(0).metadata);
-        merge.traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
+        traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
                                                            revHash, ctx, opts) {
           var rev = pos + '-' + revHash;
           if (revs.indexOf(rev) !== -1) {
