+++ lib/changes.js
@@ -1,7 +1,6 @@
 'use strict';
 var utils = require('./utils');
 var isDeleted = require('./deps/docs/isDeleted');
-var merge = require('./merge');
 var errors = require('./deps/errors');
 var EE = require('events').EventEmitter;
 var evalFilter = require('./evalFilter');
@@ -9,6 +8,8 @@ var evalView = require('./evalView');
 var parseDdocFunctionName = require('./deps/docs/parseDdocFunctionName');
 var normalizeDdocFunctionName =
   require('./deps/docs/normalizeDdocFunctionName');
+var collectLeaves = require('./deps/merge/collectLeaves');
+var collectConflicts = require('./deps/merge/collectConflicts');
 module.exports = Changes;
 utils.inherits(Changes, EE);
 
@@ -91,7 +92,7 @@ Changes.prototype.cancel = function () {
 function processChange(doc, metadata, opts) {
   var changeList = [{rev: doc._rev}];
   if (opts.style === 'all_docs') {
-    changeList = merge.collectLeaves(metadata.rev_tree)
+    changeList = collectLeaves(metadata.rev_tree)
     .map(function (x) { return {rev: x.rev}; });
   }
   var change = {
@@ -104,7 +105,7 @@ function processChange(doc, metadata, opts) {
     change.deleted = true;
   }
   if (opts.conflicts) {
-    change.doc._conflicts = merge.collectConflicts(metadata);
+    change.doc._conflicts = collectConflicts(metadata);
     if (!change.doc._conflicts.length) {
       delete change.doc._conflicts;
     }
