+++ lib/deps/ajax/prequest-browser.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var ajax = require('./ajax-core');
+var ajax = require('./ajaxCore');
 
 module.exports = function(opts, callback) {
