+++ lib/deps/ajax/prequest.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var ajax = require('./ajax-core');
+var ajax = require('./ajaxCore');
 
 module.exports = function(opts, callback) {
   // do nothing; all the action is in prerequest-browser.js
