+++ lib/deps/docs/isDeleted.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var merge = require('../../merge');
+var winningRev = require('../../deps/merge/winningRev');
 
 function getTrees(node) {
   return node.ids;
@@ -11,7 +11,7 @@ function getTrees(node) {
 //  - rev: (optional) the revision to check. defaults to winning revision
 function isDeleted(metadata, rev) {
   if (!rev) {
-    rev = merge.winningRev(metadata);
+    rev = winningRev(metadata);
   }
   var id = rev.substring(rev.indexOf('-') + 1);
   var toVisit = metadata.rev_tree.map(getTrees);
