+++ lib/deps/docs/processDocs.js
@@ -1,11 +1,11 @@
 'use strict';
 
-var merge = require('../../merge');
 var errors = require('../errors');
 var updateDoc = require('./updateDoc');
 var isDeleted = require('./isDeleted');
 var isLocalId = require('./isLocalId');
-
+var calculateWinningRev = require('../../deps/merge/winningRev');
+var merge = require('../../deps/merge');
 var collections = require('pouchdb-collections');
 var Map = collections.Map;
 
@@ -14,7 +14,7 @@ function processDocs(docInfos, api, fetchedDocs, tx, results, writeDoc, opts,
 
   function insertDoc(docInfo, resultsIdx, callback) {
     // Cant insert new deleted documents
-    var winningRev = merge.winningRev(docInfo.metadata);
+    var winningRev = calculateWinningRev(docInfo.metadata);
     var deleted = isDeleted(docInfo.metadata, winningRev);
     if ('was_delete' in opts && deleted) {
       results[resultsIdx] = errors.error(errors.MISSING_DOC, 'deleted');
@@ -85,7 +85,7 @@ function processDocs(docInfos, api, fetchedDocs, tx, results, writeDoc, opts,
           resultsIdx, docWritten, writeDoc, newEdits);
       } else {
         // Ensure stemming applies to new writes as well
-        var merged =  merge.merge([], currentDoc.metadata.rev_tree[0], 1000);
+        var merged = merge([], currentDoc.metadata.rev_tree[0], 1000);
         currentDoc.metadata.rev_tree = merged.tree;
         insertDoc(currentDoc, resultsIdx, docWritten);
       }
