+++ lib/deps/docs/updateDoc.js
@@ -1,19 +1,21 @@
 'use strict';
 
-var merge = require('../../merge');
 var errors = require('../errors');
 var isDeleted = require('./isDeleted');
 var parseDoc = require('./parseDoc').parseDoc;
+var calculateWinningRev = require('../../deps/merge/winningRev');
+var merge = require('../../deps/merge');
+var revExists = require('../../deps/merge/revExists');
 
 function updateDoc(prev, docInfo, results, i, cb, writeDoc, newEdits) {
 
-  if (merge.revExists(prev.rev_tree, docInfo.metadata.rev)) {
+  if (revExists(prev.rev_tree, docInfo.metadata.rev)) {
     results[i] = docInfo;
     return cb();
   }
 
   // sometimes this is pre-calculated. historically not always
-  var previousWinningRev = prev.winningRev || merge.winningRev(prev);
+  var previousWinningRev = prev.winningRev || calculateWinningRev(prev);
   var previouslyDeleted = 'deleted' in prev ? prev.deleted :
     isDeleted(prev, previousWinningRev);
   var deleted = 'deleted' in docInfo.metadata ? docInfo.metadata.deleted :
@@ -27,7 +29,7 @@ function updateDoc(prev, docInfo, results, i, cb, writeDoc, newEdits) {
     docInfo = parseDoc(newDoc, newEdits);
   }
 
-  var merged = merge.merge(prev.rev_tree, docInfo.metadata.rev_tree[0], 1000);
+  var merged = merge(prev.rev_tree, docInfo.metadata.rev_tree[0], 1000);
 
   var inConflict = newEdits && (((previouslyDeleted && deleted) ||
     (!previouslyDeleted && merged.conflicts !== 'new_leaf') ||
@@ -47,7 +49,7 @@ function updateDoc(prev, docInfo, results, i, cb, writeDoc, newEdits) {
   }
 
   // recalculate
-  var winningRev = merge.winningRev(docInfo.metadata);
+  var winningRev = calculateWinningRev(docInfo.metadata);
   var winningRevIsDeleted = isDeleted(docInfo.metadata, winningRev);
 
   // calculate the total number of documents that were added/removed,
