+++ lib/deps/merge/collectLeaves.js
@@ -1,0 +1,21 @@
+'use strict';
+
+var traverseRevTree = require('./traverseRevTree');
+
+function sortByPos(a, b) {
+  return a.pos - b.pos;
+}
+
+module.exports = function collectLeaves(revs) {
+  var leaves = [];
+  traverseRevTree(revs, function (isLeaf, pos, id, acc, opts) {
+    if (isLeaf) {
+      leaves.push({rev: pos + "-" + id, pos: pos, opts: opts});
+    }
+  });
+  leaves.sort(sortByPos).reverse();
+  for (var i = 0, len = leaves.length; i < len; i++) {
+    delete leaves[i].pos;
+  }
+  return leaves;
+};
