+++ lib/deps/merge/merge.js
@@ -1,0 +1,7 @@
+'use strict';
+
+
+
+var PouchMerge = {};
+
+module.exports = PouchMerge;
