+++ lib/deps/migrate.js
@@ -3,7 +3,7 @@
 var fs = require('fs');
 var path = require('path');
 var isLocalId = require('./docs/isLocalId');
-var merge = require('../merge');
+var winningRev = require('../deps/merge/winningRev');
 var levelup = require('levelup');
 var through = require('through2').obj;
 
@@ -142,7 +142,7 @@ exports.localAndMetaStores = function (db, stores, callback) {
             prefix: stores.docStore,
             type: 'del'
           });
-          var winner = merge.winningRev(ch.value);
+          var winner = winningRev(ch.value);
           Object.keys(ch.value.rev_map).forEach(function (key) {
             if (key !== 'winner') {
               this.push(formatSeq(ch.value.rev_map[key]));
