+++ lib/mapreduce/index.js
@@ -7,7 +7,7 @@ var collate = pouchCollate.collate;
 var toIndexableString = pouchCollate.toIndexableString;
 var normalizeKey = pouchCollate.normalizeKey;
 var parseIndexableString = pouchCollate.parseIndexableString;
-var createView = require('./create-view');
+var createView = require('./createView');
 var evalFunc = require('./evalfunc');
 var log;
 /* istanbul ignore else */
