+++ lib/utils.js
@@ -1,5 +1,5 @@
 /*jshint strict: false */
-var merge = require('./merge');
+var traverseRevTree = require('./deps/merge/traverseRevTree');
 exports.ajax = require('./deps/ajax/prequest');
 exports.uuid = require('./deps/uuid');
 exports.getArguments = require('argsarray');
@@ -132,7 +132,7 @@ exports.compare = function (left, right) {
 // and return a list of revs to delete
 exports.compactTree = function compactTree(metadata) {
   var revs = [];
-  merge.traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
+  traverseRevTree(metadata.rev_tree, function (isLeaf, pos,
                                                      revHash, ctx, opts) {
     if (opts.status === 'available' && !isLeaf) {
       revs.push(pos + '-' + revHash);
